refactor: move /projects/[id] components to project-view sections
